refactor: move API request types from contracts to API layer
- Move CreateAgentRequest and UpdateAgentRequest to agents.rs
- Move CreateSessionRequest and UpdateSessionRequest to sessions.rs
- Update OpenAPI schema references in main.rs
- Remove request types from contracts crate

This follows the convention established in messages.rs where API contracts
are collocated with their route handlers, keeping the contracts crate
focused on domain types (Agent, Session, Status enums).
